change the location of the cleanup function
