+++ tests/integration/local_redis_mongo.py
@@ -29,6 +29,36 @@ os.environ["DAS_USE_REDIS_CLUSTER"] = "false"
 os.environ["DAS_USE_REDIS_SSL"] = "false"
 
 
+def cleanup(request):
+    def restore_environment():
+        if DAS_MONGODB_HOSTNAME:
+            os.environ["DAS_MONGODB_HOSTNAME"] = DAS_MONGODB_HOSTNAME
+        if DAS_MONGODB_PORT:
+            os.environ["DAS_MONGODB_PORT"] = DAS_MONGODB_PORT
+        if DAS_MONGODB_USERNAME:
+            os.environ["DAS_MONGODB_USERNAME"] = DAS_MONGODB_USERNAME
+        if DAS_MONGODB_PASSWORD:
+            os.environ["DAS_MONGODB_PASSWORD"] = DAS_MONGODB_PASSWORD
+        if DAS_REDIS_HOSTNAME:
+            os.environ["DAS_REDIS_HOSTNAME"] = DAS_REDIS_HOSTNAME
+        if DAS_REDIS_PORT:
+            os.environ["DAS_REDIS_PORT"] = DAS_REDIS_PORT
+        if DAS_REDIS_USERNAME:
+            os.environ["DAS_REDIS_USERNAME"] = DAS_REDIS_USERNAME
+        if DAS_REDIS_PASSWORD:
+            os.environ["DAS_REDIS_PASSWORD"] = DAS_REDIS_PASSWORD
+        if DAS_USE_REDIS_CLUSTER:
+            os.environ["DAS_USE_REDIS_CLUSTER"] = DAS_USE_REDIS_CLUSTER
+        if DAS_USE_REDIS_SSL:
+            os.environ["DAS_USE_REDIS_SSL"] = DAS_USE_REDIS_SSL
+
+    def enforce_containers_removal():
+        _db_down()
+
+    request.addfinalizer(restore_environment)
+    request.addfinalizer(enforce_containers_removal)
+
+
 def _db_up():
     subprocess.call(
         ["bash", f"{scripts_path}/redis-up.sh", redis_port], stdout=devnull, stderr=devnull
